Update paths.c
added search dwFileAttributes for file attribute flag. This should detect more directories but some system directories are still not detected by this. it seems windows doesn't always use the directory for all directories even if it is in fact a directory. will continue to investigate but for now it works well enough.